Update - Clean Environment

+++ REST_API/rest_app.py
@@ -7,6 +7,10 @@
 import os
 import sys
 import signal
+import psutil
+
+
+# Sys Path #
 package_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.append(package_path)
 
@@ -89,17 +93,38 @@ def get_all_users_request():
         return (response, 200) if all_users_as_json is not None else (response, 500)
 
 
-@app.route('/stop_server')
-def stop_server():
+def kill_process():
+    """
+    :explanations:
+    - Kill the current process.
+
+    :return: True  - If process is still running.
+             False - If process not running anymore.
+    """
+    # Vars #
+    pid          = os.getpid()
+    process      = psutil.Process(pid)
+    process_name = process.name()
+
+    if process_name in ["python.exe", "/usr/bin/python"]:
+        os.kill(os.getpid(), signal.CTRL_C_EVENT)
+        return True
+    else:
+        return False
+
+
+@app.route('/stop_server', methods=['GET'])
+def stop_rest_api_server():
     """
     :explanations:
     - Stop running of REST API server.
 
-    :return: response (Json), 200 (Status code)
+    :return: response (Json), status_code (str)
     """
-    os.kill(os.getpid(), signal.CTRL_C_EVENT)
-    response = json.dumps({"status": "Server Stopped"})
-    return response, 200
+    if request.method == "GET":
+        is_process_killed     = kill_process()
+        response, status_code = (json.dumps({"status": "Server Stopped"}), 200) if is_process_killed is True else (json.dumps({"status": "Server Not Stopped"}), 500)
+        return response, status_code
 
 
 # Run Flask Application #
