+++ Web_Interface/web_app.py
@@ -7,13 +7,17 @@
 import os
 import sys
 import signal
+import psutil
+
+
+# Sys Path #
 package_path = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.append(package_path)
 
 
 # From #
-from flask           import Flask, make_response
-from DB.db_connector import *
+from flask             import Flask, request
+from DB.db_connector   import *
 
 
 # Create Flask Instance #
@@ -38,17 +42,38 @@ def get_user_name(user_id):
         return "<h1 id='error'>" + "No such `user id` : " + str(user_id) + "</h1>", 500
 
 
-@app.route('/stop_server')
-def stop_server():
+def kill_process():
+    """
+    :explanations:
+    - Kill the current process.
+
+    :return: True  - If process is still running.
+             False - If process not running anymore.
+    """
+    # Vars #
+    pid          = os.getpid()
+    process      = psutil.Process(pid)
+    process_name = process.name()
+
+    if process_name in ["python.exe", "/usr/bin/python"]:
+        os.kill(os.getpid(), signal.CTRL_C_EVENT)
+        return True
+    else:
+        return False
+
+
+@app.route('/stop_server', methods=['GET'])
+def stop_web_app_server():
     """
     :explanations:
     - Stop running of WEB APP server.
 
-    :return: response (Json), 200 (Status code)
+    :return: response (Json), status_code (str)
     """
-    os.kill(os.getpid(), signal.CTRL_C_EVENT)
-    response = json.dumps({"status": "Server Stopped"})
-    return response, 200
+    if request.method == "GET":
+        is_process_killed     = kill_process()
+        response, status_code = (json.dumps({"status": "Server Stopped"}), 200) if is_process_killed is True else (json.dumps({"status": "Server Not Stopped"}), 500)
+        return response, status_code
 
 
 # Run Flask Application #
